+++ CHANGELOG.md
@@ -5,9 +5,19 @@ CHANGELOG
 2019FIXME : 0.3.1 - Shuffle and Variance
 -------------------------------------------------------------------------------
 
-* shuffle an entire column with the new function : 
+__Dependencies:__
+  - tms_system_rows
+  - pgddl 0.12
+
+* dump the entire anonymized database with :
+
+  ```console
+  $ psql -q -t -A -c 'SELECT anon.dump()' the_database
+  ```
+
+* shuffle an entire column with the new function :
 	```sql
-	SELECT anon.shuffle_column('employees','salary');
+	SELECT anon.shuffle_column('employees','salary', 'id');
 	```
 
 * Add +/-33% of noise to a column with:
@@ -25,8 +35,8 @@ CHANGELOG
 * FIX #43 : Using unlogged tables was a bad idea
 
 * FIX #51 : tests & doc about explicit casting
- 
-* Add `autoload` parameter to `mask_init` function. 
+
+* Add `autoload` parameter to `mask_init` function.
   Default to TRUE for backward compatibility
 
 * Add `anon.no_extension.sql` for people in the cloud